Add inline compiler benchmark for errors
